drm: move `rawfb` drm driver into seperate module
